requirements.txt
# TOX requirements
-r requirements-tox.txt

# Minimum Django and REST framework version
Django>=4.2
djangorestframework>=3.15.0

